Extrae el fetch de comics
